functions/requirements.txt
Flask
flask-cors
requests
functions-framework
firebase-functions
Pillow
torch
torchvision
google-cloud-storage